Change default port numbers
Trying to reduce the probability of collision with dynamic ports,
while also not colliding with registered ports.
What we know about dynamic (a.k.a. ephemeral) port ranges:
- The RFC 6056 says: 1024–65535. Well, well. We are definitely not
  going under 1024.
- IANA and RFC 6335 says: 49152-65535. Which would've great, but...
- Many Linux kernels use 32768-60999,
- And some suggest setting the lower bound even lower, e.g. 15000-61000
  (https://stackoverflow.com/questions/410616/increasing-the-maximum-number-of-tcp-ip-connections-in-linux)

So apparently the best value would be below 15000. But also keep it
high to reduce the chance of colliding with somebody else's registered
port.