Do not spawn an OS thread per connection.
Using `try_send` on the `unbounded` `async_std::channel::Sender`
bridges the synchronous code inside `server.rs` and asynchronous code
in main.

The price is, for now, the dependency of `server.rs` on `async_std`,
which can be eliminated with generics or dynamic sender object/fn
but I think it's not worth it.