Make the game object Send
Initially my reasoning was:
  - Both boards use the same rules;
  - I don't want to have lifetime parameters in the board class,
    so storing a reference to the parent game class is not an
    option;
  - Rc-s are cheap;
  - Rules might be expensive and include heavy objects like an
    entire starting position.
So putting the rules into an Rc made sense. It turned out the
last part was false. All rules we have now and will have for the
foreseeable are a bunch of enums. And making the game class Send
simplifies server code. Hence removing the Rc-s.

If rules turn out to be expensive in the future, we could always
revert this and replace Rc-s with Arc-s instead.